+++ src/AdtException.ts
@@ -1,4 +1,4 @@
-import { AxiosError } from "axios"
+import { Response } from "request"
 import { AdtHTTP, session_types } from "./AdtHTTP"
 import { fullParse } from "./utilities"
 
@@ -14,7 +14,7 @@ class AdtErrorException extends Error {
     public readonly err: number,
     public readonly type: string,
     public readonly message: string,
-    public readonly parent?: AxiosError,
+    public readonly parent?: Error,
     public readonly namespace?: string,
     public readonly localizedMessage?: string
   ) {
@@ -27,10 +27,7 @@ class AdtCsrfException extends Error {
   get typeID(): symbol {
     return CSRFEXTYPEID
   }
-  constructor(
-    public readonly message: string,
-    public readonly parent: AxiosError
-  ) {
+  constructor(public readonly message: string, public readonly parent?: Error) {
     super()
   }
 }
@@ -40,7 +37,8 @@ class AdtHttpException extends Error {
     return HTTPEXTYPEID
   }
   get code() {
-    return (this.parent.response && this.parent.response.status) || 0
+    const p: any = this.parent
+    return (p.response && p.response.status) || 0
   }
   get message() {
     return this.parent.message
@@ -48,7 +46,7 @@ class AdtHttpException extends Error {
   get name() {
     return this.parent.name
   }
-  constructor(public readonly parent: AxiosError) {
+  constructor(public readonly parent: Error) {
     super()
   }
 }
@@ -58,36 +56,51 @@ export type AdtException =
   | AdtCsrfException
   | AdtHttpException
 
-export function isAdtError(e: Error): e is AdtErrorException {
+export function isAdtError(e: any): e is AdtErrorException {
   return (e as AdtErrorException).typeID === ADTEXTYPEID
 }
-export function isCsrfError(e: Error): e is AdtErrorException {
+export function isCsrfError(e: any): e is AdtErrorException {
   return (e as AdtErrorException).typeID === CSRFEXTYPEID
 }
-export function isHttpError(e: Error): e is AdtHttpException {
+export function isHttpError(e: any): e is AdtHttpException {
   return (e as AdtErrorException).typeID === HTTPEXTYPEID
 }
-export function fromException(err: AxiosError): AdtException {
+export function isAdtException(e: any): e is AdtException {
+  return isAdtError(e) || isCsrfError(e) || isHttpError(e)
+}
+const isResponse = (r: any): r is Response => !!r.statusCode
+
+export function fromException(errOrResp: Error | Response): AdtException {
+  if (isAdtException(errOrResp)) return errOrResp
   try {
-    if (!(err.response && err.response.data)) return new AdtHttpException(err)
-    if (
-      err.response.status === 403 &&
-      err.response.headers["x-csrf-token"] === "Required"
-    )
-      return new AdtCsrfException(err.response.data, err)
-    const raw = fullParse(err.response.data)
-    const root = raw["exc:exception"]
-    const getf = (base: any, idx: string) => (base ? base[idx] : "")
-    return new AdtErrorException(
-      err.response.status,
-      root.type["@_id"],
-      root.message["#text"],
-      err,
-      getf(root.namespace, "@_id"),
-      getf(root.localizedMessage, "#text")
-    )
+    if (isResponse(errOrResp)) {
+      const response: Response = errOrResp
+      if (!(response && response.body))
+        return adtException(
+          `Error ${response.statusCode}:${response.statusMessage}`
+        )
+      if (
+        response.statusCode === 403 &&
+        response.headers["x-csrf-token"] === "Required"
+      )
+        return new AdtCsrfException(response.body)
+
+      const raw = fullParse(response.body)
+      const root = raw["exc:exception"]
+      const getf = (base: any, idx: string) => (base ? base[idx] : "")
+      return new AdtErrorException(
+        response.statusCode,
+        root.type["@_id"],
+        root.message["#text"],
+        undefined,
+        getf(root.namespace, "@_id"),
+        getf(root.localizedMessage, "#text")
+      )
+    } else return new AdtHttpException(errOrResp)
   } catch (e) {
-    return new AdtHttpException(err)
+    return isResponse(errOrResp)
+      ? adtException("Unknown error in adt client")
+      : new AdtHttpException(errOrResp)
   }
 }