+++ src/api/unittest.ts
@@ -23,7 +23,7 @@ export interface UnitTestClass {
 }
 export async function runUnitTest(h: AdtHTTP, url: string) {
   const headers = { "Content-Type": "application/*", Accept: "application/*" }
-  const data = `<?xml version="1.0" encoding="UTF-8"?>
+  const body = `<?xml version="1.0" encoding="UTF-8"?>
   <aunit:runConfiguration xmlns:aunit="http://www.sap.com/adt/aunit">
   <external>
     <coverage active="false"/>
@@ -43,9 +43,9 @@ export async function runUnitTest(h: AdtHTTP, url: string) {
   const response = await h.request("/sap/bc/adt/abapunit/testruns", {
     method: "POST",
     headers,
-    data
+    body
   })
-  const raw = fullParse(response.data)
+  const raw = fullParse(response.body)
 
   const classes: UnitTestClass[] = xmlFlatArray(
     raw,